+++ src/components/home/LunchTable/index.tsx
@@ -2,7 +2,8 @@ import React from 'react';
 
 import { useTheme } from '@emotion/react';
 
-import { Text } from 'src/components';
+import { Spinner, Text } from 'src/components';
+import { useGetMealData } from 'src/hooks';
 
 import { Content } from '../Content';
 
@@ -13,17 +14,28 @@ export interface LunchTableProps {
 }
 
 export const LunchTable: React.FC<LunchTableProps> = ({ onPress }) => {
-  const theme = useTheme();
+  const { meal, isLoading, krDate } = useGetMealData();
 
+  const theme = useTheme();
   return (
     <Content icon="🍴" name="급식표" onPress={onPress}>
       <S.LunchTableTextContainer>
         <Text size={15} fontFamily="medium" color={theme.placeholder}>
           오늘의 급식
         </Text>
-        <S.LunchTableText>
-          백미밥,카레소스(추가국),등심돈까스, 망고사과샐러드,포기김치,과일음료
-        </S.LunchTableText>
+        {isLoading ? (
+          <Spinner />
+        ) : (
+          <S.LunchTableText>
+            {!isLoading && meal
+              ? meal
+                  .filter((meal) => new Date(meal.date).getDate() === krDate.getDate())
+                  .map((item) => {
+                    return item.menus.join(', ');
+                  })
+              : '불러올 급식이 없어요'}
+          </S.LunchTableText>
+        )}
       </S.LunchTableTextContainer>
     </Content>
   );
