+++ src/hooks/query/info/useGetMeal/index.tsx
@@ -16,3 +16,18 @@ export const useGetMeal = ({
     staleTime: 1000 * 60 * 60 * 24,
   });
 };
+
+export const useGetMealData = () => {
+  const KR_TIME_DIFF = 9 * 60 * 60 * 1000;
+
+  const curr = new Date();
+  const utc = curr.getTime() + curr.getTimezoneOffset() * 60 * 1000;
+  const krDate = new Date(utc + KR_TIME_DIFF);
+  const { data, isLoading } = useGetMeal({ month: `${krDate.getMonth() + 1}` });
+
+  return {
+    krDate,
+    isLoading,
+    meal: data?.data,
+  };
+};
